[REEF 708] Delete (obsolete) Delete DriverBridgeConfiguration
The removes the obsolete and unused class `DriverBridgeConfiguration`.
The file `Delete DriverBridgeConfiguration.cs` contained additional
classes which are now in their own files. Some of their methods should
be made `private`, and appropriate `[Obsolete]` annotations have been
added and JIRAs filled to be resolved after 0.13.

JIRA:
  [REEF-708](https://issues.apache.org/jira/browse/REEF-708)

Pull request:
  This closes #
